split service into faker and graphql generators
