+++ fixtures-generator/fixtures_faker/providers/product/__init__.py
