+++ fixtures-generator/graphql/responses_generator_helpers.py
